requirements.txt
alembic==1.9.2
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
greenlet==2.0.1
importlib-metadata==6.0.0
importlib-resources==5.10.2
Mako==1.2.4
MarkupSafe==2.1.1
mysqlclient==2.1.1
python-dotenv==0.21.0
pytz==2022.7.1
SQLAlchemy==1.4.46
sqlparse==0.4.3
zipp==3.11.0
